+++ package.json
@@ -21,7 +21,6 @@
   "files": [
     "vaadin-*.d.ts",
     "vaadin-*.js",
-    "@types",
     "src",
     "theme"
   ],
