+++ vaadin-crud.d.ts
@@ -1,2 +1,2 @@
 export * from './src/vaadin-crud.js';
-export * from './@types/interfaces';
+export * from './src/interfaces';
